+++ lib/src/settings.ts
@@ -1,4 +1,5 @@
 // Char used by RJSF to join the path of a field to convert it as an id
-const RJSF_ID_SEPARATOR = '_';
+export const RJSF_ID_SEPARATOR = '_';
 
-export { RJSF_ID_SEPARATOR };
+// The supported translations for the application
+export const VERNA_SUPPORTED_LOCALES = ['en-US', 'fr-FR', 'fr-CA', 'es-ES'] as const;